+++ src/error.rs
@@ -7,6 +7,12 @@ use serde_json::{json, Value};
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
+    #[error("invalid transfer: {0}")]
+    InvalidTransfer(String),
+
+    #[error("invalid form data: {0}")]
+    InvalidFormData(String),
+
     #[error("unauthorized")]
     UnauthorizedError,
 
@@ -16,41 +22,41 @@ pub enum Error {
     #[error("invalid jwt")]
     JwtInvalidError,
 
-    #[error("config error: {0:?}")]
+    #[error("config error: {0}")]
     ConfigError(#[from] config::ConfigError),
 
-    #[error("database error: {0:?}")]
+    #[error("database error: {0}")]
     DbError(#[from] r2d2::Error),
 
-    #[error("json error: {0:?}")]
+    #[error("json error: {0}")]
     JsonError(#[from] serde_json::Error),
 
-    #[error("sqlite error: {0:?}")]
+    #[error("sqlite error: {0}")]
     SQLiteError(#[from] r2d2_sqlite::rusqlite::Error),
 
-    #[error("spotify client error: {0:?}")]
+    #[error("spotify client error: {0}")]
     SpotifyClientError(#[from] rspotify::ClientError),
 
-    #[error("spotify ID error: {0:?}")]
+    #[error("spotify ID error: {0}")]
     SpotifyIdError(#[from] rspotify::model::IdError),
 
-    #[error("addr parse error: {0:?}")]
+    #[error("addr parse error: {0}")]
     AddrParseError(#[from] std::net::AddrParseError),
 
-    #[error("hmac error: {0:?}")]
+    #[error("hmac error: {0}")]
     HmacError(#[from] hmac::digest::InvalidLength),
 
-    #[error("jwt error: {0:?}")]
+    #[error("jwt error: {0}")]
     JwtError(#[from] jwt::Error),
 
-    #[error("chrono parse error: {0:?}")]
+    #[error("chrono parse error: {0}")]
     ChronoParseError(#[from] chrono::ParseError),
 
-    #[error("validation errors: {0:?}")]
-    ValidationErrors(#[from] validator::ValidationErrors),
-
-    #[error("validation error: {0:?}")]
+    #[error("validation error: {0}")]
     ValidationError(#[from] validator::ValidationError),
+
+    #[error("validation errors: {0}")]
+    ValidationErrors(#[from] validator::ValidationErrors),
 
     #[error(transparent)]
     AnyhowError(#[from] anyhow::Error),
@@ -62,6 +68,7 @@ impl IntoResponse for Error {
             Self::UnauthorizedError | Self::JwtExpiredError | Self::JwtInvalidError => {
                 (StatusCode::UNAUTHORIZED, Value::String(self.to_string()))
             }
+            Self::InvalidFormData(err) => (StatusCode::UNPROCESSABLE_ENTITY, Value::String(err)),
             Self::ValidationErrors(err) => (
                 StatusCode::UNPROCESSABLE_ENTITY,
                 json!({ "fields": err.field_errors() }),
@@ -71,7 +78,7 @@ impl IntoResponse for Error {
                 tracing::error!("{:?}", self);
                 (
                     StatusCode::INTERNAL_SERVER_ERROR,
-                    Value::String("unexpected error occurred".into()),
+                    Value::String(self.to_string()),
                 )
             }
         };