+++ src/web/view.rs
@@ -1,3 +1,4 @@
+use crate::model::{playlist::PlaylistType, watcher::Watcher};
 use askama::Template;
 use rspotify::model::SimplifiedPlaylist;
 
@@ -11,6 +12,27 @@ pub struct ConnectTemplate {
 #[template(path = "dashboard.html")]
 pub struct DashboardTemplate {
     pub name: String,
-    pub watched_playlist: Option<String>,
+    pub watchers: Vec<Watcher>,
     pub playlists: Vec<SimplifiedPlaylist>,
+}
+
+impl DashboardTemplate {
+    fn get_playlist_names(&self, watcher: &Watcher) -> (String, String) {
+        (
+            self.get_playlist_name(&watcher.from_playlist),
+            self.get_playlist_name(&watcher.to_playlist),
+        )
+    }
+
+    fn get_playlist_name(&self, playlist: &PlaylistType) -> String {
+        match playlist {
+            PlaylistType::Saved => playlist.to_string(),
+            PlaylistType::WithId(id) => self
+                .playlists
+                .iter()
+                .find(|playlist| &playlist.id.to_string() == id)
+                .map(|playlist| playlist.name.clone())
+                .unwrap_or_else(|| "(Unknown)".into()),
+        }
+    }
 }