+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Multiple watchers can now be created (and synced or removed)
+- Playlist-playlist watchers can now be created
+
+### Changed
+
+- Added better abstraction for transferring tracks between playlists
+- Renamed project from `spotify-sync` to `modulate` (a cooler name, but also required by Spotify developer terms)
+
 ## [0.3.0] - 2023-11-19
 
 ### Added